+++ Program/Optepafi/Optepafi/Models/UserModelMan/UserModels/Functionalities/IWeightComputing.cs
@@ -24,5 +24,5 @@ public interface IWeightComputing<out TTemplate, in TVertexAttributes, in TEdgeA
     /// <param name="through">Edge attributes of edge, for which is weight computed.</param>
     /// <param name="to">Vertex attributes from destination vertex.</param>
     /// <returns>Weight of edge.</returns>
-    int ComputeWeight(TVertexAttributes from, TEdgeAttributes through, TVertexAttributes to);
+    float ComputeWeight(TVertexAttributes from, TEdgeAttributes through, TVertexAttributes to);
 }
